+++ R/TM.R
@@ -3,36 +3,30 @@
 #' @title Tangent von Mises--Fisher distribution
 #'
 #' @description Density and simulation of the Tangent von Mises--Fisher (TM)
-#' distribution on
-#' \eqn{S^{p-1}:=\{\mathbf{x}\in R^p:||\mathbf{x}||=1\}}{
-#' S^{p-1} := \{x \in R^p : ||x|| = 1\}}, \eqn{p\ge 2}. The distribution arises
-#' by considering the
-#' \link[=tang-norm-decomp]{tangent-normal decomposition} with
-#' multivariate \link[=cosines-signs]{signs} distributed as a
+#' distribution on \eqn{\mathcal{S}^{p-1}:=\{\boldsymbol{x}\in
+#' \mathbb{R}^p:\|\boldsymbol{x}\|=1\}}, \eqn{p\ge 2}. The distribution arises
+#' by considering the \link[=tang-norm-decomp]{tangent-normal decomposition}
+#' with multivariate \link[=cosines-signs]{signs} distributed as a
 #' \link[=vMF]{von Mises--Fisher} distribution.
 #'
 #' @inheritParams tang-norm-decomp
-#' @param mu the directional mean \eqn{\boldsymbol{\mu}}{\mu} of the vMF
-#' used in the multivariate signs. A unit-norm vector of length \code{p - 1}.
+#' @param mu the directional mean \eqn{\boldsymbol{\mu}} of the vMF used in the
+#' multivariate signs. A unit-norm vector of length \code{p - 1}.
 #' @param kappa concentration parameter \eqn{\kappa} of the vMF used in the
 #' multivariate signs. A nonnegative scalar.
-#' @return
-#' Depending on the function:
+#' @return Depending on the function:
 #' \itemize{
 #' \item \code{d_TM}: a vector of length \code{nx} or \code{1} with the
 #' evaluated density at \code{x}.
 #' \item \code{r_TM}: a matrix of size \code{c(n, p)} with the random sample.
 #' }
-#' @details
-#' The functions are wrappers for \code{\link{d_tang_norm}} and
+#' @details The functions are wrappers for \code{\link{d_tang_norm}} and
 #' \code{\link{r_tang_norm}} with \code{d_U = \link{d_vMF}} and
 #' \code{r_U = \link{r_vMF}}.
-#' @author Eduardo García-Portugués, Davy Paindaveine, and Thomas Verdebout.
-#' @references
-#' García-Portugués, E., Paindaveine, D., Verdebout, T. (2020) On optimal tests
-#' for rotational symmetry against new classes of hyperspherical distributions.
-#' \emph{Journal of the American Statistical Association}, 115(532):1873--1887.
-#' \doi{10.1080/01621459.2019.1665527}
+#' @references García-Portugués, E., Paindaveine, D., Verdebout, T. (2020) On
+#' optimal tests for rotational symmetry against new classes of hyperspherical
+#' distributions. \emph{Journal of the American Statistical Association},
+#' 115(532):1873--1887. \doi{10.1080/01621459.2019.1665527}
 #' @examples
 #' ## Simulation and density evaluation for p = 2
 #'
@@ -47,11 +41,11 @@
 #' # Required functions
 #' r_V <- function(n) r_g_vMF(n = n, p = p, kappa = kappa_V)
 #' g_scaled <- function(t, log) {
-#'   g_vMF(t, p = p - 1, kappa = kappa_V, scaled = TRUE, log = log)
+#'   g_vMF(t, p = p, kappa = kappa_V, scaled = TRUE, log = log)
 #' }
 #'
 #' # Sample and color according to density
-#' x <- r_TM(n = n, theta = theta, r_V = r_V, mu = 1, kappa = kappa)
+#' x <- r_TM(n = n, theta = theta, r_V = r_V, mu = mu, kappa = kappa)
 #' col <- viridisLite::viridis(n)
 #' r <- runif(n, 0.95, 1.05) # Radius perturbation to improve visualization
 #' dens <- d_TM(x = x, theta = theta, g_scaled = g_scaled, mu = mu,
@@ -78,7 +72,7 @@
 #'                              pch = 16, xlab = "", ylab = "", zlab = "",
 #'                              angle = 20)
 #'
-#' ## A non-vMF angular function: g(t) = 1 - t^2. It is sssociated to the
+#' ## A non-vMF angular function: g(t) = 1 - t^2. It is associated to the
 #' ## Beta(1/2, (p + 1)/2) distribution.
 #'
 #' # Scaled angular function
@@ -104,8 +98,8 @@
 #'                              zlim = c(-1, 1), color = col[rank(dens)],
 #'                              pch = 16, xlab = "", ylab = "", zlab = "",
 #'                              angle = 20)
-#' @seealso \code{\link{tang-norm-decomp}},
-#' \code{\link{tangent-elliptical}}, \code{\link{vMF}}.
+#' @seealso \code{\link{tang-norm-decomp}}, \code{\link{tangent-elliptical}},
+#' \code{\link{vMF}}.
 #' @name tangent-vMF
 #' @aliases TM